# apps/server/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Go JWT Server
paths:
  /api/v1/auth/signup:
    post:
      summary: User signup
      description: Endpoint for user signup
      operationId: signup
      tags:
        - Auth
      parameters:
        - $ref: '#/components/parameters/UserAgent'
      requestBody:
        description: User signup request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignup'
      responses:
        '204':
          description: User created without 2FA
        '409':
          description: User already exists
        '201':
          description: User created with 2FA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpWith2FAResponse'
  /api/v1/auth/login:
    post:
      summary: User login
      description: Endpoint for user login
      operationId: login
      tags:
        - Auth
      parameters:
        - $ref: '#/components/parameters/LoginSource'
        - $ref: '#/components/parameters/UserAgent'
      requestBody:
        description: User login request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginSuccessWithJWT'
                  - $ref: '#/components/schemas/LoginRequires2FA'
                discriminator:
                  propertyName: type
...
  /api/v1/auth/refresh:
    post:
      summary: Token refresh
      description: Endpoint to get new token using refresh token
      operationId: refresh
      tags:
        - Auth
      parameters:
        - $ref: '#/components/parameters/LoginSource'
        - $ref: '#/components/parameters/UserAgent'
      requestBody:
        description: Refresh token request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Refresh'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessWithJWT'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LoginSource:
      name: x-login-source
      description: Login source
      example: web
      in: header
      required: true
      schema:
        type: string
        enum:
          - web
          - mobile
          - api
    UserAgent:
      name: User-Agent
      description: User Agent
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    UserSignup:
      type: object
      required:
        - email
        - firstName
        - lastName
        - password
        - twoFAEnabled
      properties:
        email:
          type: string
          format: email
          example: first.last@example.com
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
          minLength: 8
          description: "At least 8 characters, one uppercase, one lowercase, one number, one special character"
        twoFAEnabled:
          type: boolean
          default: false
    SignUpWith2FAResponse:
      type: object
      required:
        - qr_image
        - secret
      properties:
        qr_image:
          type: string
        secret:
          type: string
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: first.last@example.com
        password:
          type: string
...
    LoginSuccessWithJWT:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    LoginRequires2FA:
      type: object
      required:
        - type
        - temp_token
      properties:
        type:
          type: string
          enum:
            - 2fa
        temp_token:
          type: string
          description: Temporary token to use for 2FA verification
    Refresh:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
...